+++ Kernel/Sources/driver/amiga/cbm-floppy/mfm.h
@@ -11,9 +11,9 @@
 
 #include <klib/klib.h>
 
-extern void mfm_decode_sector(const uint32_t* _Nonnull input, uint32_t* _Nonnull output, int data_size);
-extern void mfm_encode_sector(const uint32_t* _Nonnull input, uint32_t* _Nonnull output, int data_size);
+extern void mfm_decode_sector(const uint32_t* _Nonnull input, uint32_t* _Nonnull output, size_t data_size);
+extern void mfm_encode_sector(const uint32_t* _Nonnull input, uint32_t* _Nonnull output, size_t data_size);
 
-extern uint32_t mfm_checksum(const uint32_t* _Nonnull input, int data_size);
+extern uint32_t mfm_checksum(const uint32_t* _Nonnull input, size_t data_size);
 
 #endif /* mfm_h */
